% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_line.R
\name{stop_line}
\alias{stop_line}
\title{stop_line}
\usage{
stop_line(line_id, destination)
}
\arguments{
\item{line_id}{is a string (the line_id)}

\item{destination}{is a string (the destination)}
}
\value{
the stops of a line
}
\description{
stop_line
}
